Restructure move feature folder structure
This change groups all controllers, fields and steps related to
the create journey into separate folders to make it easier to maintain.

It is also prep to allow the cancel journey within a move to be
expanded in a future change.
