+++ app/move/controllers/create/index.js
@@ -1,0 +1,13 @@
+const Assessment = require('./assessment')
+const Form = require('./form')
+const MoveDetails = require('./move-details')
+const PersonalDetails = require('./personal-details')
+const Save = require('./save')
+
+module.exports = {
+  Assessment,
+  Form,
+  MoveDetails,
+  PersonalDetails,
+  Save,
+}
