+++ app/move/controllers/create/personal-details.test.js
@@ -1,10 +1,10 @@
 const FormController = require('hmpo-form-wizard').Controller
 
-const Controller = require('./new.personal-details')
-const personService = require('../../../common/services/person')
-const referenceDataService = require('../../../common/services/reference-data')
-const referenceDataHelpers = require('../../../common/helpers/reference-data')
-const fieldHelpers = require('../../../common/helpers/field')
+const Controller = require('./personal-details')
+const personService = require('../../../../common/services/person')
+const referenceDataService = require('../../../../common/services/reference-data')
+const referenceDataHelpers = require('../../../../common/helpers/reference-data')
+const fieldHelpers = require('../../../../common/helpers/field')
 
 const controller = new Controller({ route: '/' })
 const genderMock = [
