+++ app/move/controllers/index.js
@@ -1,17 +1,9 @@
-const Assessment = require('./new.assessment')
-const Form = require('./new.form')
-const MoveDetails = require('./new.move-details')
-const PersonalDetails = require('./new.personal-details')
-const Save = require('./new.save')
 const cancel = require('./cancel')
+const create = require('./create')
 const view = require('./view')
 
 module.exports = {
-  Assessment,
-  Form,
-  MoveDetails,
-  PersonalDetails,
-  Save,
   cancel,
+  create,
   view,
 }
